% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdokr-package.R
\docType{package}
\name{pdokr-package}
\alias{pdokr}
\alias{pdokr-package}
\title{pdokr: Access Open Geodata from the Dutch 'PDOK' Platform}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools to discover, download, and spatially filter open geographic data from 'PDOK' (Publieke Dienstverlening Op de Kaart), the national geodata platform of the Netherlands. Datasets and their layers are searched and loaded as simple feature ('sf') objects through 'OGC' API Features endpoints, with a transparent fallback to the Web Feature Service ('WFS'), automatic pagination, and explicit coordinate reference system handling. Loaded layers can be filtered by any polygon area, and addresses or place names can be geocoded through the 'PDOK' location server. See \url{https://www.pdok.nl/} for more information about the platform and its services.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coeneisma/pdokr}
  \item \url{https://coeneisma.github.io/pdokr/}
  \item Report bugs at \url{https://github.com/coeneisma/pdokr/issues}
}

}
\author{
\strong{Maintainer}: Coen Eisma \email{coeneisma@gmail.com} (\href{https://orcid.org/0009-0007-9001-2572}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Coen Eisma \email{coeneisma@gmail.com} (\href{https://orcid.org/0009-0007-9001-2572}{ORCID}) [copyright holder]
}

}
\keyword{internal}
